refactor(tokens): shared loader core in tokens_load.go
Pure move of the agent-neutral analysis core both token commands use
(sessionTokenAnalysis, finishSessionTokenAnalysis, finishFromMetadata,
attributionDuration, metadataDuration, callWeights,
countUnmatchedSubagentRefs, applySkillEventAnchors,
assembleTokenReportView, reportAgent, modalKey/modalKeyCount,
countTokenSources, pluralHaveHas, tokenPluralSuffix) out of
checkpoint_tokens_load.go, with their tests. The checkpoint-specific
loader, legacy and task-record code stays. No behaviour change.
