+++ cmd/entire/cli/checkpoint_tokens_load_test.go
@@ -15,9 +15,6 @@ import (
 	"github.com/entireio/cli/cmd/entire/cli/tokenreport"
 )
 
-// testSkillSystematicDebugging is the skill the Claude fixture's first call runs under.
-const testSkillSystematicDebugging = "systematic-debugging"
-
 // tokenTestSession builds a stub session with the given metadata and
 // transcript for the stub reader.
 func tokenTestSession(meta checkpoint.Metadata, transcript []byte) *checkpoint.SessionContent {
@@ -347,29 +344,6 @@ func TestReadCheckpointTokenSessionsStopsWhenContextCanceled(t *testing.T) {
 	}
 }
 
-func TestApplySkillEventAnchorsLabelsUnnamedSkillLoads(t *testing.T) {
-	t.Parallel()
-
-	attribution := &types.Attribution{Calls: []types.CallUsage{
-		{Emitted: []types.ToolUseRef{{ID: "toolu_x", Tool: "Skill"}, {ID: "toolu_named", Tool: "Skill", SkillName: "keep", Detail: "keep"}}},
-		{Consumed: []types.ToolResultRef{{ToolUse: types.ToolUseRef{ID: "toolu_x", Tool: "Skill"}}}},
-	}}
-	applySkillEventAnchors(attribution, []types.SkillEvent{
-		{Skill: types.SkillEventSkill{Name: testSkillSystematicDebugging}, TranscriptAnchor: &types.SkillEventTranscriptAnchor{ToolUseID: "toolu_x"}},
-		{Skill: types.SkillEventSkill{Name: "other"}, TranscriptAnchor: &types.SkillEventTranscriptAnchor{ToolUseID: "toolu_named"}},
-	})
-	got := attribution.Calls[0].Emitted[0]
-	if got.SkillName != testSkillSystematicDebugging || got.Detail != testSkillSystematicDebugging {
-		t.Errorf("emitted ref = %+v, want the anchor's skill name", got)
-	}
-	if attribution.Calls[0].Emitted[1].SkillName != "keep" {
-		t.Errorf("a ref the attributor named must keep its name, got %+v", attribution.Calls[0].Emitted[1])
-	}
-	if attribution.Calls[1].Consumed[0].ToolUse.SkillName != testSkillSystematicDebugging {
-		t.Errorf("consumed ref = %+v", attribution.Calls[1].Consumed[0].ToolUse)
-	}
-}
-
 func TestUnmatchedSubagentNoteWordsPerAgent(t *testing.T) {
 	t.Parallel()
 
@@ -381,46 +355,6 @@ func TestUnmatchedSubagentNoteWordsPerAgent(t *testing.T) {
 	}
 	if got := unmatchedSubagentNote(agent.AgentTypeClaudeCode, 2); got != "2 subagent calls have no committed task record; that usage is not included (this backend may not store task records)." {
 		t.Errorf("claude = %q", got)
-	}
-}
-
-func TestModalKeyCountPrefersHighestThenLexical(t *testing.T) {
-	t.Parallel()
-
-	k, n := modalKeyCount(map[string]int{"gpt-5.4": 1, "claude-fable-5": 3, "claude-haiku-4-5": 3})
-	if k != "claude-fable-5" || n != 3 {
-		t.Errorf("got %q/%d, want claude-fable-5/3 (ties resolve lexically)", k, n)
-	}
-	if k, n := modalKeyCount(nil); k != "" || n != 0 {
-		t.Errorf("empty map → %q/%d", k, n)
-	}
-}
-
-// TestFinishSessionTokenAnalysis_KnownTTLPricesFiveMinuteWrites pins the
-// per-call pricing rule: an agent-written usage block with cache writes and
-// a zero 1h split is all-5m, so the class shares agree with the contributor
-// table's units and the usage table prices the row.
-func TestFinishSessionTokenAnalysis_KnownTTLPricesFiveMinuteWrites(t *testing.T) {
-	t.Parallel()
-
-	meta := &checkpoint.Metadata{SessionID: "cw-only", Agent: agent.AgentTypeClaudeCode, Model: checkpointTokensFixtureModel}
-	attribution := &types.Attribution{Calls: []types.CallUsage{
-		{Model: checkpointTokensFixtureModel, Usage: types.TokenUsage{CacheCreationTokens: 4000, APICallCount: 1}},
-		{Model: checkpointTokensFixtureModel, Usage: types.TokenUsage{CacheCreationTokens: 6000, APICallCount: 1}},
-	}}
-	a := sessionTokenAnalysis{meta: meta, attribution: attribution, efforts: map[string]int{}, models: map[string]int{}}
-	finishSessionTokenAnalysis(&a)
-
-	summed := tokenreport.SumCostShares(a.costParts...)
-	if summed.Units == 0 || summed.Units != a.attributed.PricedUnits || summed.CacheWriteUnpriced {
-		t.Fatalf("class units %v (unpriced=%v) must equal contributor units %v", summed.Units, summed.CacheWriteUnpriced, a.attributed.PricedUnits)
-	}
-
-	view := assembleTokenReportView([]sessionTokenAnalysis{a}, []*checkpoint.Metadata{meta})
-	out := renderBody(&view)
-	assertContainsAll(t, out, "Cache write, 5-minute", "(1.25×)", "10k")
-	if strings.Contains(out, "TTL not recorded") || strings.Contains(out, tokenTableUnpriced) {
-		t.Errorf("all-5m writes from a per-call block are priced, got:\n%s", out)
 	}
 }
 
@@ -471,25 +405,5 @@ func TestBuildCheckpointTokensReport_RootSummaryWhenNoSessionMetadataReadable(t 
 	assertContainsAll(t, out, "Sessions:   2", "Total", "1.7k", "of which subagents", "200", "2 session metadata files could not be read; totals are the checkpoint's root summary")
 	if strings.Contains(out, "Where it went") {
 		t.Errorf("no breakdown without readable sessions:\n%s", out)
-	}
-}
-
-func TestCountUnmatchedSubagentRefsIncludesConsumedOnlyRefs(t *testing.T) {
-	t.Parallel()
-
-	spawned := types.ToolUseRef{ID: "toolu_before", Tool: "Agent", SubagentType: "Explore"}
-	emitted := types.ToolUseRef{ID: "toolu_in", Tool: "Agent", SubagentType: "Explore"}
-	attribution := &types.Attribution{
-		Calls: []types.CallUsage{
-			{Emitted: []types.ToolUseRef{emitted}, Consumed: []types.ToolResultRef{{ToolUse: spawned}}},
-			{Consumed: []types.ToolResultRef{{ToolUse: emitted}, {ToolUse: types.ToolUseRef{ID: "toolu_bash", Tool: "Bash"}}}},
-		},
-	}
-	if got := countUnmatchedSubagentRefs(attribution); got != 2 {
-		t.Errorf("unmatched = %d, want 2 (one emitted, one consumed from before the window, each once)", got)
-	}
-	attribution.Subagents = []types.SubagentRecord{{ToolUseID: "toolu_before", SubagentType: "Explore"}}
-	if got := countUnmatchedSubagentRefs(attribution); got != 1 {
-		t.Errorf("unmatched after recording the consumed-only ref = %d, want 1", got)
 	}
 }
